archive/kas_drl/requirements.txt
# KAS DRL Requirements

# Core dependencies
torch>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# RL environment
gym>=0.26.0

# Optional for distributed training
# ray[rllib]>=2.5.0

# Optional for logging and visualization
# tensorboard>=2.13.0
# wandb>=0.15.0
# matplotlib>=3.7.0

# Optional for hyperparameter tuning
# optuna>=3.2.0

# Development dependencies
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
